Jquery updating the appointment from create.js.erb
Jquery was used to load the data from the create action. remote was
set as true which sent the data or request as JS. The create action
rendered it's view file create.js.erb. Since slim was not working for
the purpose, erb for js files were used. The JS create file changed
the current html inside of the #appointments div with the newly
queried Appointments from the create action.